feat(setup): add interactive setup wizard and indexing functionality
